// packages/mason-common/src/codegen/gql/graphql.ts
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
// ...
export type CreateInstancePayload = {
  __typename?: 'CreateInstancePayload';
  clientMutationId?: Maybe<Scalars['String']>;
  errors: Array<ErrorType>;
  instance?: Maybe<Instance>;
};
// ...
export type Customer = Node & {
  __typename?: 'Customer';
  /** Specifies whether the signed-in User can add an Alert Monitor to the Customer. */
  allowAddAlertMonitor: Scalars['Boolean'];
  /** Specifies whether the signed-in User can add a Component to the Customer. */
  allowAddComponent: Scalars['Boolean'];
  /** Specifies whether the signed-in User can add a Credential to the Customer. */
  allowAddCredential: Scalars['Boolean'];
  /** Specifies whether the signed-in User can add an Instance to the Customer. */
  allowAddInstance: Scalars['Boolean'];
  /** Specifies whether the signed-in User can add an Integration to the Customer. */
  allowAddIntegration: Scalars['Boolean'];
  /** Specifies whether the signed-in User can add a User to the Customer. */
  allowAddUser: Scalars['Boolean'];
  /** Specifies whether the signed-in User's Customer has access to legacy Credentials. */
  allowConfigureCredentials: Scalars['Boolean'];
  /** Specifies whether this Customer can use the Embedded Designer. */
  allowEmbeddedDesigner: Scalars['Boolean'];
  /** Specifies whether Instances may be enabled based on the utilization allowed by the current Plan. */
  allowEnableInstance: Scalars['Boolean'];
  /** Specifies whether Instances may be executed based on the utilization allowed by the current Plan. */
  allowExecuteInstance: Scalars['Boolean'];
  /** Specifies whether the signed-in User can manage Attachments related to this record. */
  allowManageAttachments?: Maybe<Scalars['Boolean']>;
  /** Specifies whether the signed-in User can remove the Customer. */
  allowRemove: Scalars['Boolean'];
  /** Specifies whether the signed-in User can update the Customer. */
  allowUpdate: Scalars['Boolean'];
  /** A JSON list of objects where each object has a key for name and URL that together describe the Attachment. */
  attachments?: Maybe<Array<Maybe<Attachment>>>;
  /** The URL for the avatar image. */
  avatarUrl?: Maybe<Scalars['String']>;
  /** The Customer the Component belongs to, if any. If this is NULL then the Component belongs to the Organization. */
  components: ComponentConnection;
  /** The Customer the Credential belongs to, if any. If NULL then Organization will be specified. */
  credentials: CredentialConnection;
  /** Additional notes about the Customer. */
  description: Scalars['String'];
  /** Allows for mapping an external entity to a Prismatic record. */
  externalId?: Maybe<Scalars['String']>;
  /** The ID of the object */
  id: Scalars['ID'];
  /** The Customer for which the Instance is deployed. */
  instances: InstanceConnection;
  /** The Customer the Integration belongs to, if any. If this is NULL then the Integration belongs to the Organization. */
  integrations: IntegrationConnection;
  /** The labels that are associated with the object. */
  labels?: Maybe<Array<Scalars['String']>>;
  /** The name of the Customer, which must be unique within the scope of its Organization. */
  name: Scalars['String'];
  /** The Organization to which the Customer belongs. */
  org: Organization;
  /** Indicates whether the record is starred by the signed-in User. */
  starred?: Maybe<Scalars['Boolean']>;
  /** The Customer the user belongs to, if any. If this is NULL then Organization will be specified. */
  users: UserConnection;
};
// ...
export type ErrorType = {
  __typename?: 'ErrorType';
  field: Scalars['String'];
  messages: Array<Scalars['String']>;
};
// ...
export type Instance = Node & {
  __typename?: 'Instance';
  /** Specifies whether the signed-in User can deploy the Instance. */
  allowDeploy?: Maybe<Scalars['Boolean']>;
  /** Specifies whether the signed-in User can remove the Instance. */
  allowRemove: Scalars['Boolean'];
  /** Specifies whether the signed-in User can update the Instance. */
  allowUpdate: Scalars['Boolean'];
  /** Specifies whether the signed-in User can update config variables for the Instance. */
  allowUpdateConfigVariables?: Maybe<Scalars['Boolean']>;
  /** Describes the state of configuration of this Instance. */
  configState?: Maybe<InstanceConfigState>;
  /** The Instance with which the Config Variable is associated. */
  configVariables: InstanceConfigVariableConnection;
  /** The timestamp at which the object was created. */
  createdAt: Scalars['DateTime'];
  /** The Customer for which the Instance is deployed. */
  customer: Customer;
  /** The specific version of the Instance that is deployed. */
  deployedVersion: Scalars['Int'];
  /** Additional notes about the Instance. */
  description: Scalars['String'];
  /** Specifies whether the Instance is currently enabled and in an executable state. */
  enabled: Scalars['Boolean'];
  /** The Instance for which a specific InstanceFlowConfig is being executed. */
  executionResults: InstanceExecutionResultConnection;
  /** The configuration for the IntegrationFlow associated with the Instance. */
  flowConfigs: InstanceFlowConfigConnection;
  /** The ID of the object */
  id: Scalars['ID'];
  /** Specifies whether the latest execution of this Instance resulted in a failure. */
  inFailedState: Scalars['Boolean'];
  /** The Integration that has been deployed for the Instance. */
  integration: Integration;
  /** Specifies whether the Instance can be deployed through the Marketplace. */
  isCustomerDeployable?: Maybe<Scalars['Boolean']>;
  /** Specifies whether the Instance can be upgraded through the Marketplace. */
  isCustomerUpgradeable?: Maybe<Scalars['Boolean']>;
  /** The labels that are associated with the object. */
  labels?: Maybe<Array<Scalars['String']>>;
  /** The timestamp at which the Instance was most recently deployed. */
  lastDeployedAt?: Maybe<Scalars['DateTime']>;
  /** The timestamp at which the Instance was most recently executed. */
  lastExecutedAt?: Maybe<Scalars['DateTime']>;
  /** The Instance which created the Log entry. */
  logs: LogConnection;
  /** The Instance that is being monitored by the AlertMonitor. */
  monitors: AlertMonitorConnection;
  /** The name of the Instance. */
  name: Scalars['String'];
  /** Specifies whether a deploy is needed to reflect the newest configuration for this Instance. */
  needsDeploy: Scalars['Boolean'];
  /** Indicates whether the record is starred by the signed-in User. */
  starred?: Maybe<Scalars['Boolean']>;
  /** Specifies whether the Instance has been suspended by Prismatic. */
  systemSuspended: Scalars['Boolean'];
  /** The timestamp at which the object was most recently updated.  */
  updatedAt: Scalars['DateTime'];
  /** The User Level Config variables for the requesting User on this Instance. */
  userLevelConfigVariables?: Maybe<CustomUserLevelConfigVariableConnection>;
  /** The Instance with which the User Level Config is associated. */
  userLevelConfigs: UserLevelConfigConnection;
};
// ...
export type InstanceConfigVariable = Node & {
  __typename?: 'InstanceConfigVariable';
  /** Specifies whether the signed-in User can remove the InstanceConfigVariable. */
  allowRemove: Scalars['Boolean'];
  /** Specifies whether the signed-in User can update the InstanceConfigVariable. */
  allowUpdate: Scalars['Boolean'];
  /** The Authorize URL of this Config Variable if associated with an OAuth 2.0 Connection. */
  authorizeUrl?: Maybe<Scalars['String']>;
  /** The presigned URL to fetch metadata of the content used to populate the widget, when applicable. */
  contentMetadataUrl?: Maybe<Scalars['String']>;
  /** The presigned URL to download the content used to populate the widget, when applicable. */
  contentUrl?: Maybe<Scalars['String']>;
  /** The ID of the object */
  id: Scalars['ID'];
  inputs?: Maybe<ExpressionConnection>;
  /** The Instance with which the Config Variable is associated. */
  instance: Instance;
  /** The InstanceConfigVariable which relates to the Log entry. */
  logs: LogConnection;
  /** Contains arbitrary metadata about this Config Var. */
  meta?: Maybe<Scalars['JSONString']>;
  /** The timestamp at which the OAuth2 token will automatically be refreshed, if necessary. Only applies to OAuth2 methods where refresh is necessary. */
  refreshAt?: Maybe<Scalars['DateTime']>;
  /** The Required Config Variable that is satisfied with the assignment of a Config Variable. */
  requiredConfigVariable: RequiredConfigVariable;
  /** The schedule type to show in the UI when the Config Var uses the 'schedule' dataType. */
  scheduleType?: Maybe<InstanceConfigVariableScheduleType>;
  /** Status indicating if this Connection is working as expected or encountering issues. */
  status?: Maybe<InstanceConfigVariableStatus>;
  /** The presigned URL to fetch metadata of the supplemental data that may have been fetched as part of populating the content, when applicable. */
  supplementalDataMetadataUrl?: Maybe<Scalars['String']>;
  /** The presigned URL to download supplemental data that may have been fetched as part of populating the content, when applicable. */
  supplementalDataUrl?: Maybe<Scalars['String']>;
  /** An optional timezone property for when the Config Var uses the 'schedule' dataType. */
  timeZone?: Maybe<Scalars['String']>;
  /** The value for the Required Config Variable that becomes part of the Instance definition. */
  value?: Maybe<Scalars['String']>;
};
// ...
export type InstanceConfigVariableConnection = {
  __typename?: 'InstanceConfigVariableConnection';
  /** List of edges containing the nodes in this connection. */
  edges: Array<Maybe<InstanceConfigVariableEdge>>;
  /** List of nodes in this connection. */
  nodes: Array<Maybe<InstanceConfigVariable>>;
  /** Pagination data for this connection. */
  pageInfo: PageInfo;
  /** Total count of nodes available. */
  totalCount: Scalars['Int'];
};
// ...
export enum InstanceConfigVariableStatus {
  /** active */
  Active = 'ACTIVE',
  /** error */
  Error = 'ERROR',
  /** pending */
  Pending = 'PENDING'
}
// ...
export enum RequiredConfigVariableDataType {
  /** Boolean */
  Boolean = 'BOOLEAN',
  /** Code */
  Code = 'CODE',
  /** Connection */
  Connection = 'CONNECTION',
  /** Credential */
  Credential = 'CREDENTIAL',
  /** Date */
  Date = 'DATE',
  /** Jsonform */
  Jsonform = 'JSONFORM',
  /** Number */
  Number = 'NUMBER',
  /** Objectfieldmap */
  Objectfieldmap = 'OBJECTFIELDMAP',
  /** Objectselection */
  Objectselection = 'OBJECTSELECTION',
  /** Picklist */
  Picklist = 'PICKLIST',
  /** Schedule */
  Schedule = 'SCHEDULE',
  /** String */
  String = 'STRING',
  /** Timestamp */
  Timestamp = 'TIMESTAMP'
}
// ...
export type CreateInstanceMutationVariables = Exact<{
  integrationId: Scalars['ID'];
  customerId: Scalars['ID'];
  name: Scalars['String'];
  description?: InputMaybe<Scalars['String']>;
  labels?: InputMaybe<Array<InputMaybe<Scalars['String']>> | InputMaybe<Scalars['String']>>;
  configVariables?: InputMaybe<Array<InputMaybe<InputInstanceConfigVariable>> | InputMaybe<InputInstanceConfigVariable>>;
}>;


export type CreateInstanceMutation = { __typename?: 'RootMutation', createInstance?: { __typename: 'CreateInstancePayload', instance?: { __typename: 'Instance', id: string, name: string, description: string, labels?: Array<string> | null, customer: { __typename: 'Customer', id: string, name: string }, integration: { __typename: 'Integration', id: string, name: string, versionNumber: number }, configVariables: { __typename?: 'InstanceConfigVariableConnection', nodes: Array<{ __typename?: 'InstanceConfigVariable', id: string, authorizeUrl?: string | null, status?: InstanceConfigVariableStatus | null, value?: string | null, requiredConfigVariable: { __typename?: 'RequiredConfigVariable', dataType: RequiredConfigVariableDataType } } | null> } } | null, errors: Array<{ __typename: 'ErrorType', field: string, messages: Array<string> }> } | null };